refactor: rename useStats→useViews, update endpoint to /api/views
